+++ milkman/src/main/java/milkman/ui/main/HotkeyManager.java
@@ -6,6 +6,8 @@ import javafx.scene.Scene;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyCodeCombination;
 import javafx.scene.input.KeyCombination;
+import milkman.ui.commands.AppCommand;
+import milkman.ui.commands.AppCommand.EditCurrentEnvironment;
 import milkman.ui.commands.UiCommand;
 import milkman.ui.commands.UiCommand.NewRequest;
 import milkman.ui.commands.UiCommand.RenameActiveRequest;
@@ -17,6 +19,7 @@ import milkman.utils.Event;
 @Singleton
 public class HotkeyManager {
 
+	public final Event<AppCommand> onAppCommand = new Event<AppCommand>();
 	public final Event<UiCommand> onCommand = new Event<UiCommand>();
 	
 	public void registerGlobalHotkeys(Scene scene) {
@@ -31,6 +34,10 @@ public class HotkeyManager {
 		
 		scene.getAccelerators().put(new KeyCodeCombination(KeyCode.S, KeyCombination.CONTROL_DOWN), 
 				() -> onCommand.invoke(new SaveActiveRequest()));
+		
+
+		scene.getAccelerators().put(new KeyCodeCombination(KeyCode.E, KeyCombination.CONTROL_DOWN), 
+				() -> onAppCommand.invoke(new EditCurrentEnvironment()));
 	}